Switched state variable from 16 bit to 8 bit

+++ verbalEyes_speed_controller.c
@@ -15,7 +15,7 @@
 
 #define TAB "    "
 
-static uint16_t state = 0;
+static uint8_t state = 0;
 static size_t timeout = 0;
